app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Arturo Cuicas <arturo.cuicas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
pydantic = "^2.6.4"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
